improve chat and types

+++ models/product_crud.go
@@ -6,8 +6,8 @@ type DeleteProductById struct {
 
 type ProductUpdateModel struct {
 	ProductName   string  `json:"product_name"`
-	Weight        int     `json:"weight"`
-	Calories      int     `json:"calories"`
+	Weight        float64 `json:"weight"`
+	Calories      float64 `json:"calories"`
 	Proteins      float64 `json:"proteins"`
 	Fats          float64 `json:"fats"`
 	Carbohydrates float64 `json:"carbohydrates"`
